MA-17076 plat: imx8: Only save data section for cold reboot
In some cases, the bl31 won't be reloaded when spl is not supported,
commit 17de039 adds the save/restore data section to fix boot issues
which is caused by the dirty data in data section of previous boot.

However, sometimes the backup data section in dram won't be erased
totally in board cold reboot, it will be restored and modify the
'correct' data section which will cause the board hang.

This commit uses a global flag 'data_section_restore_flag' which is
initialized as '0x1' and should be stored in data section to indicate
the save/restore behavior.

Test: cold/warm reboot on imx8qm/imx8qxp.

Signed-off-by: Ji Luo <ji.luo@nxp.com>