LF-3635 gic: add new function that clears all ITs routed to a cluster
This new function is used by imx8 when a partition is rebooting.
when GIC distributor is already configured (means that one OS
has already booted), we clear all the interrupts routed to the
cluster currently (re-)executing ATF, we change active and pending
IRQs into inactive before clearing them. This avoids some old
interrupts to disturb the kernel boot. The interrupt source is
assumed to be reset or clock-gated to prevent it from generating
new IRQs.

Signed-off-by: Seb Fagard <sebastien.fagard@nxp.com>
Signed-off-by: Fabrice Goucem <fabrice.goucem@nxp.com>
Signed-off-by: Nitin Garg <nitin.garg@nxp.com>
Signed-off-by: Abel Vesa <abel.vesa@nxp.com>