MA-21624 Remove active setting SNVS_LPCR_LPTA_EN | SNVS_LPCR_LPWUI_EN
Currently code actively reads the SNVS_LPCR register. From the kernel if
there is an alarm present, the alarm will not be disabled. Therefore
SNVS_LPCR_LPTA_EN and SNVS_LPCR_LPWUI_EN are not needed.

If active setting is present, there is a possibility that shutting down the
machine after suspend/resume will result in a reboot. Because the
alarm is turned off on resume if only one alarm set by healthd, the
active setting enables the alarm on shutdown in ATF, this causes
the alarm to work thus leading to reboot.

Test:
echo +600 > /sys/class/rtc/rtc0/wakealarm
Shutdown after several suspend/resume, the machine will reboot after
600s of setting the alarm.

Change-Id: Ie96e87029559c1640d714ebf5626c6c1df561eb2