LF-11278 imx95: Ensure the cluster can be powered off in suspend.
Fix incorrect save/restore of GIC context
Enable NOC(with OCRAM retention) and wakeup MIX to be powered off
in system suspend state.
Add support for informing the SM on the core to be resumed. This is
used as workaround for the HW bug: all wakeup sources are directed to
the A55 cluster. The SM will resume the core specified as the core to
be woken up on a wakeup interrupt.

Signed-off-by: Ranjani Vaidyanathan <ranjani.vaidyanathan@nxp.com>