imx8mp: Add DRAM2_MAP region to memory mapping
For the boards with DRAM size larger than 3GB, the rest of the DRAM
is not mapped.

From docs/plat/imx8m.rst:

'DRAM mapping is done with MT_RW attributes, as it is required for
data exchange between EL3 and EL1&0 software.
The reason for the MT_RW attribute mapping scheme is the fact that the SMC
API to get the status and events is called from NS world passing destination
pointers which are located in DRAM. Mapping DRAM without MT_RW permissions
causes those locations not to be filled, which in turn causing EL1&0 software
not to receive replies.'