spd: trusty: Add check for LATE_MAPPED_BL32 before mapping the first page of bl32
The memory mapping of the first page of bl32 before peeking into Trusty
code to see if we have a 32 or 64 bit image has been implemented for a particular
case when LATE_MAPPED_BL32 is set to 1. Other than this case, bl32 is already mapped
and if we don't check for LATE_MAPPED_BL32, we will get an error on mmap_add_region_check:
region completely overlaps another.

Add also check for PLAT_XLAT_TABLES_DYNAMIC, because mmap_add_dynamic_region
and mmap_remove_dynamic_region function definitions are added on condition
that PLAT_XLAT_TABLES_DYNAMIC set to 1.

Signed-off-by: Natalia Kovalenko <natasha.k@variscite.com>