LF-16386 fix(imx93): correct the system resume flow
The AP core will check if the M33 core is active or not during the
suspend & resume flow to do different system setting. In some corner
case, the M33 core status may not be the same when ap in the susepnd
and resume flow, it will lead to wrong system config in resume flow.

Add a new flag to make sure the suspend/resume flow is correct. if
the flag is true, when doing resume we should reverse all the system
config that is done in suspend flow.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>